+++ .babelrc
@@ -1,6 +1,9 @@
 {
   "presets": [
-    "@babel/preset-env",
+    ["@babel/preset-env", {
+      "modules": false,
+      "loose": true
+    }],
     "@babel/preset-flow"
   ]
 }
